+++ dont_fret/web/state.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import solara
 
 from dont_fret.config import cfg
-from dont_fret.web.models import BurstColorList
-from dont_fret.web.new_models import FRETStore, ListStore, ThreadedDataManager
+from dont_fret.web.models import BurstColorList, FRETStore, ListStore
+from dont_fret.web.new_models import ThreadedDataManager
 from dont_fret.web.reactive import BurstSettingsReactive, SnackbarReactive
 
 APP_TITLE = "Don't FRET!"
